Reorganize presentation helpers and add desktop window constraints
- Extract desktop window sizing to Platforms/Desktop/DesktopWindowHelper.cs
- Add minimum window size (450x640) via OverlappedPresenter.PreferredMinimum*
- Move text helpers to Presentation/Common/Text/
- Move practice-specific helpers to Presentation/Practice/Common/
- Move FrozenHeaderTable to Presentation/Grammar/
- Update imports across all affected pages

+++ PaliPractice/PaliPractice/Presentation/Common/LayoutConstants.cs
@@ -19,7 +19,7 @@ public static class LayoutConstants
 {
     #region Width/Height Constraints
 
-    public static double ContentMaxWidth => 450;
+    public static int ContentMaxWidth => 450;
 
     // Translation block width as percentage of card width
     public static double TranslationWidthRatio => 0.76;
